refact: move code to specific dialect
